+++ AmitX/Kernel/include/app.h
@@ -1,0 +1,7 @@
+#ifndef APP_H
+#define APP_H
+
+void task_counter();
+void task_heartbeat();
+
+#endif